+++ satella/coding/decorators.py
@@ -35,6 +35,8 @@ def wraps(cls_to_wrap: tp.Type) -> tp.Callable[[tp.Type], tp.Type]:
             cls.__name__ = cls_to_wrap.__name__
         if hasattr(cls_to_wrap, '__module__'):
             cls.__module__ = cls_to_wrap.__module__
+        if hasattr(cls_to_wrap, '__annotations__'):
+            cls.__annotations__ = cls_to_wrap.__annotations__
         return cls
 
     return outer