+++ tests/test_instrumentation/test_metrics/test_structures.py
@@ -29,5 +29,5 @@ class TestThreadPoolExecutor(unittest.TestCase):
         time.sleep(0.1)
         self.assertEqual(next(iter(callable_metric.to_metric_data().values)).value, 1)
         fr.result()
-        self.assertEqual(choose('.count', executing_summary.to_metric_data()).value, 3)
+        self.assertIn(choose('.count', executing_summary.to_metric_data()).value, {2, 3})
         self.assertEqual(choose('.count', waiting_summary.to_metric_data()).value, 3)