+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.7.3
 
+* fixed `wraps` to provide __annotations__ as well
 * a slight fix for PyPy
 
 # v2.7.3